+++ Tools/notesTools.py
@@ -2,6 +2,7 @@ import Globals as gs
 import re
 from Tools.errorHandler import ErrorLogging
 from Tools.netTools import SendWSMessage, OSCOrder, ExecuteOSCList, SendOSCMessage, ResetRoundOSC, NewRound
+from Tools.Items.Encounters import TerrorID8Pages
 import asyncio
 
 # Check the content of a log and parse all data
@@ -72,9 +73,16 @@ def ParseContent(i_content, i_fileName, i_cursor):
                                 if args[3] != '':       # I found a case with an 8 pages where the game returned an empty round with other killers before the round ended
                                     gs.roundKiller = f"{args[0]} {args[1]} {args[2]}"
                                     gs.roundType = args[3]
-                                    OSCorderList.append(OSCOrder('round_killer1', gs.oscJsonProfile['round_killer1']['variable'], int(args[0])))
-                                    OSCorderList.append(OSCOrder('round_killer2', gs.oscJsonProfile['round_killer2']['variable'], int(args[1])))
-                                    OSCorderList.append(OSCOrder('round_killer3', gs.oscJsonProfile['round_killer3']['variable'], int(args[2])))
+                                    if args[3] != '8 Pages':
+                                        OSCorderList.append(OSCOrder('round_killer1', gs.oscJsonProfile['round_killer1']['variable'], int(args[0])))
+                                        OSCorderList.append(OSCOrder('round_killer2', gs.oscJsonProfile['round_killer2']['variable'], int(args[1])))
+                                        OSCorderList.append(OSCOrder('round_killer3', gs.oscJsonProfile['round_killer3']['variable'], int(args[2])))
+                                    else:
+                                        # In case of 8 pages what value goes into each killer number for OSC changes
+                                        result = TerrorID8Pages(args[0], args[1], args[3])
+                                        OSCorderList.append(OSCOrder('round_killer1', gs.oscJsonProfile['round_killer1']['variable'], result))
+                                        OSCorderList.append(OSCOrder('round_killer2', gs.oscJsonProfile['round_killer2']['variable'], int(args[1])))
+                                        OSCorderList.append(OSCOrder('round_killer3', gs.oscJsonProfile['round_killer3']['variable'], int(args[0])))
                                     OSCorderList.append(OSCOrder('round_type', gs.oscJsonProfile['round_type']['variable'], gs.oscJsonProfile['round_type']['values'][args[3]]))
                             case "round_unknown":
                                 print('round_unknown placeholder')