+++ CHANGELOG.md
@@ -7,10 +7,13 @@
 - Removed `codes-folder` from config.xml.default
 - Cleaned up code and fixed references to the Codes folder.
 - Moved control.xml from xml to json format. [#63](https://github.com/69MichelleDB/ToNCodes/issues/63)
+- Changed default OSC profile to Kittenji's OSC standard.
 
 **Fixes**
 - Fixes to prevent null cursors when parsing Log files. [#63](https://github.com/69MichelleDB/ToNCodes/issues/63)
 - Refresh encounter data if the file is missing on load. [#62](https://github.com/69MichelleDB/ToNCodes/issues/62)
+- For OSC calls, if the round is 8 pages it'll send the ID correctly following Kittenji's standard. [#61](https://github.com/69MichelleDB/ToNCodes/issues/61)
+- Fixed crash when enabling OSC on first boot.
 
 ## [alpha-0.7.9] - 2025.04.11